:recycle: Refactor blend mode and fills into their own submodules
