+++ render-wasm/src/math.rs
@@ -1,4 +1,3 @@
 use skia_safe as skia;
 
 pub type Rect = skia::Rect;
-pub type Matrix = skia::Matrix;
